Renaming of actions, fix user relation for enquiry, test changed
