+++ app/Models/Enquiry.php
@@ -53,7 +53,7 @@ class Enquiry extends Model
      */
     public function user(): BelongsTo
     {
-        return $this->belongsTo(User::class, 'id', 'user_id');
+        return $this->belongsTo(User::class, 'user_id', 'id');
     }
 
     /**
